+++ src/ui/App/useRemoteBundle.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import { Bundle as BundleData } from "../../audioset";
-import { fetchAudioset } from "../../player/fetchAudioset";
+import { fetchAudioset } from "./fetchAudioset";
 
 export function useRemoteBundle(idOrUrl: string) {
   const [bundle, setBundle] = useState<BundleData | null>(null);
